+++ src/main/main.java
@@ -1,6 +1,7 @@
 package main;
 
 import DNA.*;
+import Info.Info;
 
 public class main {
 
@@ -9,7 +10,7 @@ public class main {
 		//System.out.println(nuc1);
 		String str1 = "aaattttggggcccc";
 		//System.out.println(str1.length());
-		Fragment frag1 = new Fragment("aaaattttggggcccc", 0);
+		Fragment frag1 = new Fragment("aaaattttggggcccc", (short)0);
 		System.out.println(frag1);
 
 	}
